Testing a way of processing audit information
This audit is to test the viability of processing counts for information
on the clients end. We first need to convert the audit to a web friendly
format such as JSON and then compute the counts for chain. It may be the
case that we should preprocess the JSON files so all chains include only
the final link.