Initial stage for implementing search
This commit isn't yet at baseline but it lays a very minimal foundation
for testing and implementing a working search functionality. The upcoming
changes planned will allow for processing audit information and
having multiple filters simultaneously.

Some things to note here is that we're passing around course selections
as an object of the entire course. Doing comparisons don't work for some
unknown reason as what I assume to be identical objects are being seen as
unique to javascript. The workaround for this is to use the course_code
as the unique bit rather than letting javascript deal with it. Avoid
using comparisons for these objects and manually do it instead.