Implement rudimentary search
This commit introduces a working search with a ranking system to
prioritize criteria such as a keyword matching the title taking
precedance over the description.

The search is designed such that we would be able to apply multiple
filters at once which is important for flexibility in the future. In
addition to the search functionality, the display of course listing is
now animated and interative allowing you to select and add a course to
the schedule.