Improve search results UI
This commit introduces a new format for the course cards which is a
table design containing the search results. It is much cleaner and
provides all information that the SOC site would. It currently doesn't
display the course pre-requisites, counts for and is lacking animations.

To ensure flexibility and improve the UI/UX, it's better to implement a
table manually using CSS grids.