Implemented grid table for search
The search results are now displayed using a grid rather than a table
but there is an issue with animations. They still don't work. The
culprit here is that to render the flip animation we need an entire div
to be moved but you can't easily wrap the table rows in a div since it
doesn't preserve the grid structure, the workaround is to add a display
property set to contents but this property means that the wrapper isn't
considered when applying the animate directive. Hoping this saves
someone else from countless debugging hours.