Fix Search Results animation
One of sveltes requirements is that an element which has the animate
directive applied to it needs to be an intermediate to a keyed block,
this restricts you if you're using components to separate different
parts of the application.

Normally a standard wrapper over the component would work but in this
case it isn't quite simple since we're simulating a table structure
using CSS grid for the search results table. We have to preserve the
structure while applying a wrapper and as mentioned in previous commits
we cannot preserve the structure without applying the display: contents
property which leads to the animation directive not working.

The workaround here is to have each card be its own grid by wrapping it
with a grid similar to that of the table header and then applying the
animate to this wrapper. This maintains the structure of aligning the
cards to the header and solves our problem.

It is important to note that the design of the cards here are
temporary.