Implemented Timetable View
This commit adds the critical timetable view, it leverages an existing
library and overrides the parts which are not necessary. A future
implementation should consider implementing a calendar from scratch.

This commit also adjusts the logic of the ScheduleCard so that the
schedule is loaded for a given card rather than the search feature. This
makes it easier to process whether a selected course exists within the
schedule for error handling. It might make sense to have some sort of
global variable which loads the selected schedule rather than having
each card load it.