Implemented course conflict logic
This commit adds functionality to filter whether courses from the search
conflict with those which are selected. It works as expected but a
current bug is that for courses such as english 100 where you have
multiple valid sections, if you conflict with any one of these than the
entire course is filtered out but this should not be the case.