Fix conflicting courses logic
This commit builds on the previous one fixing the described error. We
can now account for special cases such as lecture and recitation
pairs when filtering. We also account for cases where a course has
multiple possible sections such as 76100. A conflict with one section
shouldn't rule out the entire course.