Implement a working Timetable
This commit introduces a working timetable which updates as you select
and change the selected sections of a course. They update accordingly
and each course displayed has a predictable color based of the course
code. Due to using an external library I've had to override some
options such as its CSS as the week view is set to the current week
which means we need to calculate which days a course would have occured
in this week.

The library also doesn't support repeating events or recording the same
event on multiple days which means we need to create an event entry for
each day of the course.

Currently this system works but is buggy in two ways, you can't hover
over the course to see its information and if the course is about 50
minutes long the title and course code is cutoff. The other bug is the
colors are too similar and need to be adjusted.