Establish audit format and extract audit tree
This commit introduces a new format for audits sent to the frontend.
The format is as follows, you have a json file containing an array where
the first index is the unique list of requirements for that audit,
followed by each course object which contains the course code,
requirement, inclusion/exclusion, type.

eg.
[
  [
    "req 1",
    "req2"
  ],
  {
    "Course or code": "eg",
    "Requirement": "eg",
    "Inclusion/Exclusion": "eg",
    "Type": "Course"
  },
  ...
]