Add counts for to cards and fix major bugs
This commit adds one of the most important features which is displaying
a courses counts for data. In addition to this, certain filtering
features are now enabled and selected courses section is refined.

Some majors bugs are also fixed where audit information would display
once but on subsequent refreshes would crash. The root of this issue is
metadata was being added and the database would return the transformed
object. Audit information also wasn't loaded due to a missing await :0
keyword.

Work needs to be done to ensure all promises are dealt with properly and
ensure better error handling when requests fail. Currently the site just
crashes and some promises are left unhandled.