Fix reactivity when changing schedule
This commit introduces a fix for the timetable not refreshing if the
search component isn't rendered. It also fixes cases where if a schedule
is changed and a selected course is not on that schedule then the
timetable crashes and the selectcard doesn't accurately reflect that the
selected course isn't on the schedule. When changing schedule if the
sections differ we also weren't accounting for that but we now display
the correct information to the user.

Added footer with a link to the github repo and also the license.