Improve search perf and capability, added DaisyUI
This commit now improves some of the search performance by using a set
rather than an array wherever possible. Similar changes will be made for
other operations. In addition, we can now filter out cleared courses.

We introduce DaisyUI for this update as is part of the transition to a
usable beta interface. DaisyUI is compatible with the tailwind UI kit
and adds to it with semantic colors and prebuilt themes/styling.