Improved UI and added new features
This commit now introduces a more polished UI where daisyUI is used in
certain instances, namely the buttons and the navigation bar. Part of
the new sugar is a new LoadingOverlay component which will now be used
throughout the program where necessary.

I've also added deletion of schedules which are currently stored,
polished the mechanism for saving schedules to ensure we always serve
the latest one. Finally, schedule names can be renamed, say goodbye to
Plan A, Plan B and hello to more meaningful names :)