Rework search UI and fix select card crash
This commit introduces a major reworking of the search UI creating an
overlay so users don't need to scroll up and scroll down constantly. It
fixes the search controls as well.

The select card was also buggy where you could only select a section
after refreshing the page. Currently, there is a bug where if you change
schedule, the select card doesn't update with the new details. The logic
behind how a schedule change affects the application needs to be thought
about more.