Fix search checkbox bug, Add search cache
This update fixes the checkbox bug whereby if you search and select the
checkbox, reentering search would reflect the incorrect state where the
box may ticked but not shown to the user.

Added search cache as searching for conflicting courses takes time and
it only needs to be run if we have a new course added or if we're
filtering for the first time.

Fix bug where require in tailwind config wouldn't work, renaming to
common js prevents ES module warning and ensures file is interpreted
correctly. On LTS versions of node it would still work without this
fix. The pleasure of using a rolling distribution :)